rename Life to Main
